fix php8.1
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@13950 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ includes/classes/validpass.php
@@ -236,8 +236,8 @@ class validpass {
 		$itoa64 = './ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
 
 		$output = '$2a$';
-		$output .= chr(ord('0') + $this->iteration_count_log2 / 10);
-		$output .= chr(ord('0') + $this->iteration_count_log2 % 10);
+		$output .= chr((int)(ord('0') + $this->iteration_count_log2 / 10));
+		$output .= chr((int)(ord('0') + $this->iteration_count_log2 % 10));
 		$output .= '$';
 
 		$i = 0;
